+++ moht/start.py
@@ -34,12 +34,11 @@ def run_tk():
 
 def run_qt():
     """Function to start Mod Helper Tool QtGUI."""
-    from PyQt5 import QtCore
-    from PyQt5.QtCore import QLibraryInfo, QTranslator, QLocale
+    from PyQt5.QtCore import Qt, QCoreApplication, QLibraryInfo, QLocale, QTranslator
     from PyQt5.QtWidgets import QApplication
     from moht.qtgui import MohtQtGui
 
-    QtCore.QCoreApplication.setAttribute(QtCore.Qt.AA_ShareOpenGLContexts)
+    QCoreApplication.setAttribute(Qt.AA_ShareOpenGLContexts)
     signal.signal(signal.SIGTERM, signal.SIG_DFL)
     signal.signal(signal.SIGINT, signal.SIG_DFL)
     app = QApplication(sys.argv)
@@ -48,7 +47,7 @@ def run_qt():
     if translator.load(QLocale.system(), 'qtbase', '_', QLibraryInfo.location(QLibraryInfo.TranslationsPath)):
         app.installTranslator(translator)
     translator = QTranslator(app)
-    if translator.load(QLocale.system(), 'qtgui', '-', here(__file__)):  # change to _
+    if translator.load(f':translations/i18n/qtgui_{QLocale.system().name()}.qm'):
         app.installTranslator(translator)
 
     try: