+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # v0.7.0
 * make `tes3cmd-v0.40` default version (it not require `perl-Config-IniFiles`)
-* remove Report button from main window
+* QtGUI:
+  * nice report in separate tab
+  * remove Report button from main window
 
 # v0.6.2
 * Fix missing internal field for TkGui