+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   * nice report in separate tab
   * remove Report button from main window
   * move close and update button to common area 
+  * double click on report item, copied path of plugin to clipboard
 
 # v0.6.2
 * Fix missing internal field for TkGui